Rename Equation and BlockEquation
