+++ src/components/blocks/Equation.svelte
