+++ src/components/blocks/components/KatexEquation.svelte
